% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao.R
\name{encontrar_pior_ano_time}
\alias{encontrar_pior_ano_time}
\title{FUNCAO ENCONTRAR PIOR ANO DE UM DETERMINADO TIME}
\usage{
encontrar_pior_ano_time(timeaux)
}
\arguments{
\item{timeaux}{time do brasileirão}
}
\value{
Retorna uma tibble
}
\description{
A funcao tem como objetivo encontrar o pior ano de um determinado time do brasileirao
}
\examples{
encontrar_pior_ano_time("Botafogo")
}
