% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-r.R
\docType{data}
\name{partidas_brasileirao}
\alias{partidas_brasileirao}
\title{PARTIDAS BRASILEIRAO.}
\format{
As variaveis da base sao
\describe{
\item{Temporada}{Descrição da coluna 1}
\item{Data}{Descrição da coluna 2}
\item{time_casa}{Descrição da coluna 2}
\item{gols}{Descrição da coluna 2}
...
}
}
\source{
"https://github.com/williamorim/brasileirao"
}
\usage{
partidas_brasileirao
}
\description{
Partidas do brasileirao de 2003 a 2020.
}
\keyword{datasets}
